Refactor WebView Architecture: Complete modular redesign
Features:
- Modular component-based architecture
- Dependency injection pattern
- Template system with token replacement
- Organized CSS with VS Code theme integration

Structure:
- src/webview/components/ - Reusable UI components
- src/webview/helpers/ - Utility classes
- src/webview/styles/ - Modular CSS files
- src/webview/templates/ - HTML templates
- src/webview/scripts/ - Client-side JavaScript

Impact:
- Reduced main provider from 789 to ~170 lines
- Improved maintainability and testability
- Enhanced code organization and readability
- Zero breaking changes - full backward compatibility

Technical:
- WebviewBuilder pattern for content generation
- Event delegation for better performance
- Clean separation of HTML, CSS, JS, and TS
- SOLID principles applied throughout"
