+++ src/webview/README.md
@@ -1,0 +1,148 @@
+# ClickCMD WebView Architecture
+
+## 📁 Estrutura Refatorada
+
+A arquitetura do WebView foi completamente refatorada seguindo as melhores práticas de organização e separação de responsabilidades.
+
+```
+src/
+├── webview/
+│   ├── components/           # Componentes reutilizáveis
+│   │   ├── headerComponent.ts
+│   │   └── commandRowComponent.ts
+│   ├── helpers/              # Classes auxiliares
+│   │   ├── iconHelper.ts     # Gerenciamento de ícones SVG
+│   │   ├── htmlHelper.ts     # Utilitários HTML (escape, nonce)
+│   │   ├── styleHelper.ts    # Gerenciamento de estilos CSS
+│   │   └── templateHelper.ts # Sistema de templates
+│   ├── scripts/              # Scripts do WebView
+│   │   └── webviewScript.ts  # Lógica JavaScript do cliente
+│   ├── styles/               # Estilos CSS modulares
+│   │   ├── variables.css     # Variáveis CSS do VS Code
+│   │   ├── base.css          # Estilos base
+│   │   ├── buttons.css       # Componentes de botão
+│   │   ├── header.css        # Estilos do cabeçalho
+│   │   ├── command-row.css   # Estilos das linhas de comando
+│   │   └── responsive.css    # Media queries responsivas
+│   ├── templates/            # Templates HTML
+│   │   ├── main.html         # Template principal
+│   │   └── error.html        # Template de erro
+│   └── webviewBuilder.ts     # Construtor principal do WebView
+```
+
+## 🏗️ Componentes Principais
+
+### WebviewBuilder
+- **Responsabilidade**: Orquestrar a construção completa do WebView
+- **Funcionalidades**: 
+  - Integrar todos os componentes
+  - Gerenciar templates e estilos
+  - Construir HTML final
+
+### HeaderComponent
+- **Responsabilidade**: Gerar o cabeçalho da interface
+- **Funcionalidades**:
+  - Botões de controle (Toggle Terminal, Edit, Refresh, Clear Changes)
+  - Logo e título da aplicação
+  - Layout responsivo
+
+### CommandRowComponent
+- **Responsabilidade**: Gerar linhas de comando individuais
+- **Funcionalidades**:
+  - Exibição de comandos com preview
+  - Botões de ação (Edit, Clear, Run)
+  - Campo de input para overrides temporários
+  - Estado visual para comandos modificados
+
+## 🎨 Sistema de Estilos
+
+### Modularização CSS
+- **variables.css**: Variáveis CSS do VS Code para integração com tema
+- **base.css**: Estilos fundamentais e layout
+- **buttons.css**: Todos os estilos de botões e estados
+- **header.css**: Layout e estilos do cabeçalho
+- **command-row.css**: Estilos específicos dos comandos
+- **responsive.css**: Media queries para diferentes tamanhos de tela
+
+### Vantagens
+- ✅ Manutenção mais fácil
+- ✅ Reutilização de código
+- ✅ Separação clara de responsabilidades
+- ✅ Facilita testes unitários
+- ✅ Melhor organização do código
+
+## 🔧 Helpers
+
+### IconHelper
+- Centraliza todos os ícones SVG
+- Facilita adição/modificação de ícones
+- Padronização visual
+
+### HtmlHelper
+- Escape de HTML para segurança
+- Geração de nonce para CSP
+- Utilitários de manipulação HTML
+
+### StyleHelper
+- Carregamento e concatenação de estilos CSS
+- Integração com sistema de temas do VS Code
+- Otimização de CSS
+
+### TemplateHelper
+- Sistema de templates com substituição de tokens
+- Carregamento de templates HTML
+- Flexibilidade para múltiplos layouts
+
+## 📱 JavaScript Modular
+
+### WebviewScript
+- **Event Delegation**: Usa delegação de eventos para melhor performance
+- **Separação de Concerns**: Cada função tem responsabilidade específica
+- **Maintainability**: Código mais limpo e organizável
+
+### Melhorias Implementadas
+- ✅ Event delegation em vez de múltiplos event listeners
+- ✅ Funções específicas para cada tipo de evento
+- ✅ Melhor tratamento de erros
+- ✅ Código mais legível e maintível
+
+## 🚀 Benefícios da Refatoração
+
+### Para Desenvolvedores
+- **Código mais limpo**: Cada arquivo tem uma responsabilidade clara
+- **Facilidade de manutenção**: Alterações são localizadas
+- **Reutilização**: Componentes podem ser reutilizados
+- **Testes**: Mais fácil testar componentes isolados
+
+### Para Performance
+- **CSS modular**: Apenas estilos necessários são carregados
+- **Event delegation**: Melhor performance com menos listeners
+- **Código otimizado**: Estrutura mais eficiente
+
+### Para Escalabilidade
+- **Arquitetura extensível**: Novos componentes podem ser adicionados facilmente
+- **Padrões consistentes**: Estrutura padronizada para crescimento
+- **Manutenibilidade**: Código organizado para equipes maiores
+
+## 📋 Migration Notes
+
+### Antes (Monolítico)
+```typescript
+// Tudo em CommandsViewProvider
+private generateWebviewContent() {
+  // 500+ linhas de HTML/CSS/JS inline
+}
+```
+
+### Depois (Modular)
+```typescript
+// Arquitetura distribuída
+const webviewBuilder = new WebviewBuilder(context);
+const html = webviewBuilder.generateMainContent(config, effective, webview);
+```
+
+### Compatibilidade
+- ✅ Mantém toda funcionalidade existente
+- ✅ Zero breaking changes na API
+- ✅ Mesma interface do usuário
+- ✅ Performance melhorada
