+++ src/webview/styles/variables.css
@@ -1,0 +1,22 @@
+/* CSS Variables for VS Code Theme Integration */
+:root {
+	/* VSCode Theme Variables */
+	--vscode-font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, sans-serif;
+	--base-font-size: max(11px, calc(var(--vscode-editor-font-size, 14px) - 2px));
+	--vscode-foreground: var(--vscode-foreground);
+	--vscode-editor-background: var(--vscode-editor-background);
+	--vscode-panel-border: var(--vscode-panel-border);
+	--vscode-list-hoverBackground: var(--vscode-list-hoverBackground);
+	--vscode-button-background: var(--vscode-button-background);
+	--vscode-button-foreground: var(--vscode-button-foreground);
+	--vscode-button-hoverBackground: var(--vscode-button-hoverBackground);
+	--vscode-button-secondaryBackground: var(--vscode-button-secondaryBackground);
+	--vscode-button-secondaryForeground: var(--vscode-button-secondaryForeground);
+	--vscode-input-border: var(--vscode-input-border);
+	--vscode-input-background: var(--vscode-input-background);
+	--vscode-input-foreground: var(--vscode-input-foreground);
+	--vscode-focusBorder: var(--vscode-focusBorder);
+	--vscode-descriptionForeground: var(--vscode-descriptionForeground);
+	--vscode-textCodeBlock-background: var(--vscode-textCodeBlock-background);
+	--vscode-editor-font-family: var(--vscode-editor-font-family);
+}
