+++ src/webview/templates/error.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<meta charset="UTF-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<title>ClickCMD - Error</title>
+	<style>
+		body { 
+			font-family: var(--vscode-font-family); 
+			padding: 20px; 
+			color: var(--vscode-foreground); 
+			background: var(--vscode-editor-background);
+		}
+		.error { color: var(--vscode-errorForeground); }
+	</style>
+</head>
+<body>
+	<div class="error">
+		<h3>⚠️ Error</h3>
+		<p>Failed to load configuration. Please check the .cmmds file and try again.</p>
+	</div>
+</body>
+</html>
