+++ src/webview/templates/main.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<meta charset="UTF-8">
+	<meta http-equiv="Content-Security-Policy" content="default-src 'none'; style-src 'unsafe-inline' {{cspSource}}; script-src 'nonce-{{nonce}}';">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<title>ClickCMD</title>
+	{{styles}}
+</head>
+<body>
+	<div class="container">
+		{{header}}
+		<main class="content">
+			{{content}}
+		</main>
+	</div>
+	{{script}}
+</body>
+</html>
